+++ tsdoc/src/parser/__tests__/NodeParserCode.test.ts
@@ -23,3 +23,69 @@ test('01 Code span basic, negative', () => {
     ' */'
   ].join('\n'));
 });
+
+test('03 Code fence, positive', () => {
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * This is a code fence with all parts:',
+    ' * ```a language!   ',
+    ' *   some `code` here',
+    ' * ```   ',
+    ' */'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * This is a code fence with no language or trailing whitespace:',
+    ' * ```',
+    ' *   some `code` here',
+    ' * ```*/'
+  ].join('\n'));
+});
+
+test('04 Code fence, negative', () => {
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Code fence incorrectly indented:',
+    ' *    ```',
+    ' */'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Code fence not starting the line:',
+    ' *a```',
+    ' */'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Code fence not being terminated 1:',
+    ' * ```*/'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Code fence not being terminated 2:',
+    ' * ``` some stuff',
+    ' */'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Language having backticks:',
+    ' * ``` some stuff ```',
+    ' */'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Closing delimiter being indented:',
+    ' * ```',
+    ' * code',
+    ' *      ```',
+    ' */'
+  ].join('\n'));
+  TestHelpers.parseAndMatchNodeParserSnapshot([
+    '/**',
+    ' * Closing delimiter not being on a line by itself:',
+    ' * ```',
+    ' * code',
+    ' * ```  a',
+    ' */'
+  ].join('\n'));
+});